third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/attr-lang-inherit-expected.txt

CONSOLE ERROR: line 55: Uncaught TypeError: host.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL attr-lang-inherit Uncaught TypeError: host.createShadowRoot is not a function
Harness: the test ran to completion.
